Simplify/refactor handling of `main`/`_start`
This change is NFC for non-standalone mode.

For standalone more we no longer have any special handling for `main`.
This is because `_start` is the entry symbol for standalone mode.  We no
longer include `_main` in EXPORTED_FUNCTIONS by default.  With this
change `--no-entry` is only way build a reactor.  EXPORTED_FUNCTIONS
no longer effects whether we build a reactor or a command.